Changed price to bigdecimal to align with currency best practices

+++ src/main/kotlin/market/LimitOrder.kt
@@ -6,13 +6,13 @@ import java.util.*
 data class LimitOrder(
     val side: String,
     val quantity: BigDecimal,
-    val price: Double,
+    val price: BigDecimal,
     val pair: String,
     val orderTimestamp: Long = System.currentTimeMillis(),
     val orderId: UUID = UUID.randomUUID()
 )
 
 data class CompletedOrder(
-    val quantity: BigDecimal, val price: Double, val side: String,
+    val quantity: BigDecimal, val price: BigDecimal, val side: String,
     val orderId: UUID, val completedTimeStamp: Long = System.currentTimeMillis()
 )
